Relocates and Fixes path saving tests
